feat: RacingCarApplication 수정 및 racingcar 패키지 아래로 이동
